refactor(agent): extract context-graph resolution into mixin holder
Move existence/curation checks, sync-request envelope parse/build/auth,
on-chain participant + curator resolution, wire-id<->local-id mapping,
access-policy resolution, private-CG participant listing and the
listContextGraphs enumeration (~1.3k lines) into dkg-agent-cg-resolve.ts.
